requirements.txt
matplotlib
numpy
pandas
PyQt5
openpyxl
tabula-py
scipy
scikit-posthocs
xlrd
outlier_utils